refactor: Moved node types into thier own directory
