+++ CMakeLists.txt
@@ -13,24 +13,23 @@ set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${CMAKE_SOURCE_DIR}/cmake/")
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
 set(PUBLIC_HEADERS
-	include/moth_ui/animation/animation_clip_controller.h                                                                                                                                         
-	include/moth_ui/animation/animation_clip.h                                                                                                                                                    
-	include/moth_ui/animation/animation_controller.h                                                                                                                                              
-	include/moth_ui/animation/animation_event.h                                                                                                                                                   
-	include/moth_ui/animation/animation_track_controller.h                                                                                                                                        
-	include/moth_ui/animation/animation_track.h                                                                                                                                                   
-	include/moth_ui/animation/clip_controller.h                                                                                                                                                   
-	include/moth_ui/animation/keyframe.h                                                                                                                                                          
+	include/moth_ui/animation/animation_clip_controller.h
+	include/moth_ui/animation/animation_clip.h
+	include/moth_ui/animation/animation_controller.h
+	include/moth_ui/animation/animation_event.h
+	include/moth_ui/animation/animation_track_controller.h
+	include/moth_ui/animation/animation_track.h
+	include/moth_ui/animation/clip_controller.h
+	include/moth_ui/animation/keyframe.h
 	include/moth_ui/blend_mode.h
 	include/moth_ui/context.h
-	include/moth_ui/event_dispatch.h
-	include/moth_ui/event_listener.h
 	include/moth_ui/events/event_animation.h
+	include/moth_ui/events/event_dispatch.h
 	include/moth_ui/events/event.h
 	include/moth_ui/events/event_key.h
+	include/moth_ui/events/event_listener.h
 	include/moth_ui/events/event_mouse.h
 	include/moth_ui/font_factory.h
-	include/moth_ui/group.h
 	include/moth_ui/ifont_factory.h
 	include/moth_ui/ifont.h
 	include/moth_ui/iimage_factory.h
@@ -51,12 +50,14 @@ set(PUBLIC_HEADERS
 	include/moth_ui/layout/layout_entity_type.h
 	include/moth_ui/layout/layout.h
 	include/moth_ui/layout/layout_rect.h
-	include/moth_ui/node_clip.h
 	include/moth_ui/node_factory.h
-	include/moth_ui/node.h
-	include/moth_ui/node_image.h
-	include/moth_ui/node_rect.h
-	include/moth_ui/node_text.h
+	include/moth_ui/nodes/group.h
+	include/moth_ui/nodes/node_clip.h
+	include/moth_ui/nodes/node.h
+	include/moth_ui/nodes/node_image.h
+	include/moth_ui/nodes/node_rect.h
+	include/moth_ui/nodes/node_text.h
+	include/moth_ui/nodes/widget.h
 	include/moth_ui/text_alignment.h
 	include/moth_ui/ui_fwd.h
 	include/moth_ui/utils/color.h
@@ -67,7 +68,6 @@ set(PUBLIC_HEADERS
 	include/moth_ui/utils/serialize_utils.h
 	include/moth_ui/utils/vector.h
 	include/moth_ui/utils/vector_serialization.h
-	include/moth_ui/widget.h
 )
 set(SOURCES ${SOURCES}
 	src/animation/animation_clip_controller.cpp
@@ -78,7 +78,6 @@ set(SOURCES ${SOURCES}
 	src/animation/keyframe.cpp
 	src/context.cpp
 	src/font_factory.cpp
-	src/group.cpp
 	src/layers/layer.cpp
 	src/layers/layer_stack.cpp
 	src/layout/layout_cache.cpp
@@ -90,12 +89,13 @@ set(SOURCES ${SOURCES}
 	src/layout/layout_entity_rect.cpp
 	src/layout/layout_entity_ref.cpp
 	src/layout/layout_entity_text.cpp
-	src/node_clip.cpp
-	src/node.cpp
 	src/node_factory.cpp
-	src/node_image.cpp
-	src/node_rect.cpp
-	src/node_text.cpp
+	src/nodes/group.cpp
+	src/nodes/node_clip.cpp
+	src/nodes/node.cpp
+	src/nodes/node_image.cpp
+	src/nodes/node_rect.cpp
+	src/nodes/node_text.cpp
 )
 
 find_package(nlohmann_json REQUIRED)
