+++ include/moth_ui/nodes/node_rect.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "moth_ui/node.h"
+#include "moth_ui/nodes/node.h"
 
 #include <memory>
