+++ src/animation/animation_controller.cpp
@@ -2,7 +2,7 @@
 #include "moth_ui/animation/animation_controller.h"
 #include "moth_ui/animation/animation_track_controller.h"
 #include "moth_ui/layout/layout_entity.h"
-#include "moth_ui/node.h"
+#include "moth_ui/nodes/node.h"
 
 namespace {
     using namespace moth_ui;
