+++ src/layout/layout.cpp
@@ -1,7 +1,7 @@
 #include "common.h"
 #include "moth_ui/layout/layout.h"
 #include "moth_ui/animation/animation_clip.h"
-#include "moth_ui/group.h"
+#include "moth_ui/nodes/group.h"
 #include "moth_ui/node_factory.h"
 #include "moth_ui/context.h"
