+++ src/layout/layout_entity_clip.cpp
@@ -1,6 +1,6 @@
 #include "common.h"
 #include "moth_ui/layout/layout_entity_clip.h"
-#include "moth_ui/node_clip.h"
+#include "moth_ui/nodes/node_clip.h"
 
 namespace moth_ui {
     LayoutEntityClip::LayoutEntityClip(LayoutRect const& initialBounds)
