+++ src/layout/layout_entity_image.cpp
@@ -1,6 +1,6 @@
 #include "common.h"
 #include "moth_ui/layout/layout_entity_image.h"
-#include "moth_ui/node_image.h"
+#include "moth_ui/nodes/node_image.h"
 
 namespace moth_ui {
     LayoutEntityImage::LayoutEntityImage(LayoutRect const& initialBounds)
