+++ src/layout/layout_entity_rect.cpp
@@ -1,6 +1,6 @@
 #include "common.h"
 #include "moth_ui/layout/layout_entity_rect.h"
-#include "moth_ui/node_rect.h"
+#include "moth_ui/nodes/node_rect.h"
 
 namespace moth_ui {
     LayoutEntityRect::LayoutEntityRect(LayoutRect const& initialBounds)
