+++ src/layout/layout_entity_ref.cpp
@@ -1,6 +1,6 @@
 #include "common.h"
 #include "moth_ui/layout/layout_entity_ref.h"
-#include "moth_ui/group.h"
+#include "moth_ui/nodes/group.h"
 #include "moth_ui/layout/layout.h"
 #include "moth_ui/animation/animation_clip.h"
