+++ src/layout/layout_entity_text.cpp
@@ -1,6 +1,6 @@
 #include "common.h"
 #include "moth_ui/layout/layout_entity_text.h"
-#include "moth_ui/node_text.h"
+#include "moth_ui/nodes/node_text.h"
 
 namespace moth_ui {
     LayoutEntityText::LayoutEntityText(LayoutRect const& initialBounds)
