+++ src/nodes/group.cpp
@@ -1,9 +1,9 @@
 #include "common.h"
-#include "moth_ui/group.h"
+#include "moth_ui/nodes/group.h"
 #include "moth_ui/layout/layout_entity_group.h"
 #include "moth_ui/animation/animation_clip.h"
 #include "moth_ui/events/event_dispatch.h"
-#include "moth_ui/node_clip.h"
+#include "moth_ui/nodes/node_clip.h"
 #include "moth_ui/context.h"
 #include "moth_ui/node_factory.h"
 #include "moth_ui/events/event_animation.h"
