+++ src/nodes/node_clip.cpp
@@ -1,5 +1,5 @@
 #include "common.h"
-#include "moth_ui/node_clip.h"
+#include "moth_ui/nodes/node_clip.h"
 #include "moth_ui/layout/layout_entity_clip.h"
 
 namespace moth_ui {
