+++ src/nodes/node_image.cpp
@@ -1,5 +1,5 @@
 #include "common.h"
-#include "moth_ui/node_image.h"
+#include "moth_ui/nodes/node_image.h"
 #include "moth_ui/layout/layout_entity.h"
 #include "moth_ui/layout/layout_entity_image.h"
 #include "moth_ui/context.h"
