+++ src/nodes/node_rect.cpp
@@ -1,5 +1,5 @@
 #include "common.h"
-#include "moth_ui/node_rect.h"
+#include "moth_ui/nodes/node_rect.h"
 #include "moth_ui/layout/layout_entity_rect.h"
 #include "moth_ui/context.h"
