+++ src/nodes/node_text.cpp
@@ -1,5 +1,5 @@
 #include "common.h"
-#include "moth_ui/node_text.h"
+#include "moth_ui/nodes/node_text.h"
 #include "moth_ui/layout/layout_entity.h"
 #include "moth_ui/layout/layout_entity_text.h"
 #include "moth_ui/context.h"
